improve test lasyout
